feat(navigation): refactor navigation components for improved structure and mobile support
- Added ChplDesktopNav for desktop navigation with CMS ID Creator and Compare Products features.
- Introduced ChplMobileNavDrawer for mobile navigation, including collapsible sections for CMS ID Creator, Compare Products, Resources, and Shortcuts.
- Updated navigation-top to integrate new desktop and mobile navigation components, enhancing responsiveness.

[#OCD-4307-UI]
