--- src/app/navigation/navigation.scss
@@ -1,244 +1,0 @@
-@import "~bootstrap-sass/assets/stylesheets/bootstrap/_variables.scss";
-
-.default-visible {
-    display: block !important;
-
-    &.toggle-visible {
-        display: none !important;
-    }
-}
-
-.default-hidden {
-    display: none !important;
-
-    &.toggle-visible {
-        display: block !important;
-    }
-}
-
-.navbar {
-    transition: all .4s ease;
-}
-
-.navbar-inverse .navbar-nav>li>a {
-    font-family: "Lato", sans-serif;
-    color: white;
-    text-decoration: none;
-}
-
-.navbar-inverse .navbar-nav>li>a:hover {
-    text-decoration: underline;
-}
-
-.navbar-inverse {
-    background: #001439;
-    box-shadow: none;
-    min-height: 100px;
-}
-
-.navbar-inverse .navbar-nav>.active>a,
-.navbar-inverse .navbar-nav>.active>a:focus,
-.navbar-inverse .navbar-nav>.active>a:hover,
-.navbar-inverse .navbar-nav>.open>a {
-    background: #0070c9;
-    background-image: none;
-}
-
-.navbar-brand img {
-    transition: all .4s ease;
-}
-
-.navbar-inverse .navbar-nav>.active>a,
-.navbar-inverse .navbar-nav>.open>a {
-    box-shadow: none;
-}
-
-.navbar-scrolled {
-    @media (min-width: 767px) {
-        min-height: 60px !important;
-
-        .container {
-            min-height: 60px !important;
-        }
-
-        .navbar-brand {
-            height: 60px !important;
-
-            img {
-                max-height: 32px !important;
-                margin: 0 !important;
-            }
-        }
-    }
-}
-
-.navbar-admin {
-    background: #fff;
-
-    .navbar-nav {
-
-        >li {
-
-            .btn-link {
-                color: #000;
-            }
-
-            button.admin-active {
-                border-bottom: #012a68 solid thick;
-                border-radius: 0;
-            }
-
-            >a {
-                color: #000;
-            }
-        }
-    }
-}
-
-@media (max-width: 767px) {
-
-    .navbar-fixed-top {
-        position: relative;
-        top: auto;
-    }
-
-    .navbar-fixed-bottom {
-        position: relative;
-    }
-
-    .navbar-nav > li > a {
-        line-height: 20px;
-        padding-top: 0;
-        padding-bottom: 10px;
-    }
-}
-
-@media (min-width: 767px) {
-
-    .navbar-fixed-top .container {
-        position: relative;
-        min-height: 100px;
-        overflow: visible;
-    }
-
-    .navbar-fixed-top .nav-main {
-        position: absolute;
-        top: 60%;
-        right: 15px;
-        transform: translateY(-50%);
-        white-space: nowrap;
-    }
-
-    /* Vertically center the brand/logo to align with nav items */
-    .navbar-inverse .navbar-brand {
-        height: 100px;
-        padding-top: 0;
-        padding-bottom: 0;
-        display: -ms-flexbox;
-        display: -webkit-box;
-        display: flex;
-        -ms-flex-align: center;
-        -webkit-box-align: center;
-        align-items: center;
-    }
-}
-
-.navigation-shown {
-    @media (max-width: 767px) {
-        margin-top: 0;
-    }
-
-    @media (min-width: 767px) {
-        margin-top: 132px;
-    }
-}
-
-.navigation-hidden {
-    @media (max-width: 767px) {
-        margin-top: 0;
-    }
-
-    @media (min-width: 767px) {
-        margin-top: 100px;
-    }
-}
-
-.toggle-navigation-open {
-    transition: all .4s ease;
-    position: absolute;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    z-index: -1;
-
-    .btn-link {
-        color: #fff;
-    }
-}
-
-.toggle-navigation-closed {
-    position: relative;
-    top: 0;
-    background-color: #001439;
-
-    .btn-link,
-    .btn-link:focus,
-    .btn-link:hover {
-        color: #fff;
-    }
-}
-
-.navbar-scrolled {
-
-    .toggle-navigation-open {
-        //bottom: -30px;
-        bottom: 0;
-    }
-}
-
-.nav {
-
-    .dropdown-menu {
-        padding: 10px;
-        border: 3px solid #0070c9;
-
-        p {
-            margin-left: 2px;
-            margin-right: 2px;
-        }
-
-        li a {
-            font-weight: bold;
-            text-decoration: none;
-        }
-
-        li a:hover {
-            text-decoration: underline;
-        }
-
-        &#compare-widget-dropdown {
-            max-height: 64vh;
-            overflow-y: auto;
-        }
-
-        &#widget-dropdown {
-            max-height: 64vh;
-            overflow-y: auto;
-            min-width: $screen-xs * 4 / 5;
-        }
-    }
-    .spaced-out {
-        margin: 10px 40px;
-    }
-
-    @media (min-width: 768px) {
-
-        .navbar-right {
-            margin-right: 0;
-        }
-    }
-}
-
-.announcement.badge {
-    white-space: inherit;
-}
